move inference req/resp types to a separate file. Future annotations messes up pydantic
