+++ llm/inference/__init__.py
@@ -1,5 +1,6 @@
 from llm.inference.base import InferenceEngine
 from llm.inference.multiproc import MultiprocessEngine
 from llm.inference.transformer import TransformersEngine
+from llm.inference.types import InferenceRequest, InferenceResponse
 from llm.inference.utils import LogitsProcessorConfig
 from llm.inference.vllm_client import VLLMClient
