+++ llm/inference/transformer.py
@@ -1,9 +1,7 @@
 from __future__ import annotations
-from dataclasses import dataclass, field
 import gc
 
 from typing import Any, Dict, Iterable, List, Optional, Set, Tuple, Union
-from uuid import uuid4
 
 import torch
 from torch.types import Device
@@ -12,7 +10,8 @@ from transformers import (
     PreTrainedTokenizerBase,
 )
 from llm.inference.base import InferenceEngine
-from llm.inference.utils import DataclassBase, LogitsProcessorConfig, check_stop_str
+from llm.inference.types import InferenceRequest, InferenceResponse, InferenceState
+from llm.inference.utils import LogitsProcessorConfig, check_stop_str
 from llm.model_configs import ModelConfig
 from llm.utils.logs import get_logger
 
@@ -113,7 +112,7 @@ class TransformersEngine(InferenceEngine):
         echo_prompt: bool = False,
         stop_token_ids: Optional[List[int]] = None,
         stop_strings: Union[str, List[str]] = "",
-        stream_interval: int = 2,
+        token_interval: int = 2,
         **kwargs,
     ) -> Iterable[str]:
         """
@@ -125,7 +124,7 @@ class TransformersEngine(InferenceEngine):
             echo_prompt=echo_prompt,
             stop_token_ids=stop_token_ids,
             stop_strings=stop_strings,
-            stream_interval=stream_interval,
+            token_interval=token_interval,
             **kwargs
         )
         for response in self.generate_batch_stream([request]):
@@ -345,12 +344,12 @@ class TransformersEngine(InferenceEngine):
         """
         Decode tokens and check for stopping conditions
         """
-        stream_interval = state.req.stream_interval
-        if not state.resp.stopped and not state.req.stop_strings and stream_interval <= 0:
+        token_interval = state.req.token_interval
+        if not state.resp.stopped and not state.req.stop_strings and token_interval <= 0:
             # Skip decoding tokens when not needed
             return
 
-        if (stream_interval > 0 and state.resp.tokens_generated % stream_interval == 0) or state.resp.stopped:
+        if (token_interval > 0 and state.resp.tokens_generated % token_interval == 0) or state.resp.stopped:
             # Decode tokens, and check if an update needs to be yielded
             if state.req.echo_prompt:
                 rfind_start = len(state.prompt)
@@ -468,85 +467,6 @@ class TransformersEngine(InferenceEngine):
         self.encoder_attn_cache.clear()
         gc.collect()
         torch.cuda.empty_cache()
-
-
-@dataclass
-class InferenceRequest(DataclassBase):
-    """
-    Input data for inference
-    """
-    prompt: Union[str, List[int]]
-
-    uid: str = field(default_factory=lambda: uuid4().hex)
-    max_new_tokens: int = 256
-    stream_interval: int = 2
-    echo_prompt: bool = False
-    stop_strings: Union[str, List[str]] = ""
-    stop_token_ids: Optional[List[int]] = None
-
-    temperature: float = 1.0
-    top_p: float = 1.0
-    top_k: int = -1
-    repetition_penalty: float = 1.0
-    do_sampling: Optional[bool] = None
-
-
-@dataclass
-class InferenceResponse(DataclassBase):
-    """
-    Ouput data from inference.
-    """
-    uid: str
-    output: str = ""
-    stopped: bool = False
-    stopped_reason: str = ""
-    tokens_generated: int = 0
-
-
-@dataclass
-class InferenceState:
-    """
-    Stores information about an in-progress inference request/response
-    """
-    req: InferenceRequest
-    prompt: str
-    input_ids: List[int]
-    output_updated: bool = False
-
-    resp: InferenceResponse = field(init=False)
-    tokens: List[int] = field(init=False)
-    logits_config: LogitsProcessorConfig = field(init=False)
-
-    def __post_init__(self):
-        self.resp = InferenceResponse(self.req.uid)
-        if self.req.echo_prompt:
-            self.tokens = list(self.input_ids)
-        else:
-            self.tokens = []
-        self.logits_config = LogitsProcessorConfig(
-            temperature=self.req.temperature,
-            top_p=self.req.top_p,
-            top_k = self.req.top_k,
-            repetition_penalty=self.req.repetition_penalty,
-            do_sampling=self.req.do_sampling,
-        )
-
-    def add_token(self, token: int):
-        self.tokens.append(token)
-        self.resp.tokens_generated += 1
-        is_stop_token = self.req.stop_token_ids and (token in self.req.stop_token_ids)
-        if is_stop_token or self.resp.tokens_generated == self.req.max_new_tokens:
-            self.set_stopped(
-                "stop token" if is_stop_token else "max new tokens"
-            )
-
-    def set_stopped(self, reason: str):
-        self.resp.stopped = True
-        self.resp.stopped_reason = reason
-
-    def set_output(self, output: str):
-        self.resp.output = output
-        self.output_updated = True
 
 
 class PastKVCache:
