+++ llm/qa/parser.py
@@ -8,7 +8,7 @@ from datasets import Dataset
 from langchain.embeddings.base import Embeddings
 from langchain.text_splitter import TextSplitter
 
-from llm.utils.enum import StrEnum
+from llm.utils.types import StrEnum
 
 logger = logging.getLogger(__name__)
